Fixed server crash per #1197 due to loading language files on the server.
