+++ src/main/java/minecrafttransportsimulator/jsondefs/JSONConfigSettings.java
@@ -90,6 +90,7 @@ public class JSONConfigSettings{
 		public String comment2 = "Inside each entry exist values for the potency of the fluids that can be considered to be that fuel type.";
 		public String comment3 = "You can make engines take different types of fluids for their fuels by editing these values.  See the in-game handbook for more information.";
 		public String comment4 = "If you are having troubles getting the name for a particular fluid, run the game with the mods you want and check the 'lastLoadedFluids' list.  This contains all the fluids that were present during the last run of the game.";
+		public String comment5 = "The first entry in the list is the registered fluid name, while the second entry will be the displayed name.  Only use the first name, as the second name changes based on game language settings.";
 		public Map<String, Map<String, Double>> fuels;
 		public Map<String, String> lastLoadedFluids;
